requirements.txt
streamlit==1.28.1
supabase==2.3.5
anthropic==0.7.11
openai==1.3.9
google-generativeai==0.3.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
